Resolve "Add ProjectCodeQuality Entity"

See merge request abms/products/rise8/midas/midas-api!99

+++ src/main/java/mil/af/abms/midas/api/project/dto/UpdateProjectDTO.java
@@ -23,7 +23,7 @@ public class UpdateProjectDTO implements Serializable {
     @UniqueName(isNew = false)
     private String name;
 
-    private Long gitlabProjectId;
+    private Integer gitlabProjectId;
 
     @TeamExists(allowNull = true)
     private Long teamId;
